+++ apps/meteor/server/startup/watchDb.ts
@@ -1,4 +1,4 @@
-import { api } from '@rocket.chat/core-services';
+import { api, dbWatchersDisabled } from '@rocket.chat/core-services';
 import { Logger } from '@rocket.chat/logger';
 import { MongoInternals } from 'meteor/mongo';
 
@@ -19,12 +19,17 @@ watcher.watch().catch((err: Error) => {
 	process.exit(1);
 });
 
-setInterval(function _checkDatabaseWatcher() {
-	if (watcher.isLastDocDelayed()) {
-		SystemLogger.error('No real time data received recently');
-	}
-}, 20000);
+if (!dbWatchersDisabled) {
+	setInterval(function _checkDatabaseWatcher() {
+		if (watcher.isLastDocDelayed()) {
+			SystemLogger.error('No real time data received recently');
+		}
+	}, 20000);
+}
 
 export function isLastDocDelayed(): boolean {
+	if (dbWatchersDisabled) {
+		return true;
+	}
 	return watcher.isLastDocDelayed();
 }